pyproject.toml
[project]
name = "rag_telegram_bot"
version = "0.1.0"
description = "RAG Telegram Bot with LM Studio"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "tiktoken>=0.7.0",
    "python-telegram-bot>=21.0",
    "nest-asyncio>=1.6.0",
    "jupyterlab>=4.0.0",
]
